Add Implementation Delegate for the Old Architecture
